Fix cyclic dependencies.
