+++ coreblocks/utils/fifo.py
@@ -1,6 +1,5 @@
 from amaranth import *
-from transactron import Method, def_method, Priority, TModule
-from transactron._utils import MethodLayout
+from transactron import Method, def_method, Priority, TModule, MethodLayout
 from coreblocks.utils._typing import ValueLike
