+++ transactron/core/sugar.py
@@ -1,15 +1,19 @@
 from amaranth import *
-from typing import Callable, Optional
+from typing import Callable, Optional, TYPE_CHECKING
 from .modules import TModule
-from .method import Method
 from .typing import ValueLike, RecordDict
 from .._utils import method_def_helper
 from coreblocks.utils import assign, AssignType
+
 __all__ = [
     "def_method",
 ]
 
-def def_method(m: TModule, method: Method, ready: ValueLike = C(1)):
+if TYPE_CHECKING:
+    from .method import Method
+
+
+def def_method(m: TModule, method: "Method", ready: ValueLike = C(1)):
     """Define a method.
 
     This decorator allows to define transactional methods in an
